requirements.txt
geopandas>=0.12.2
pandas>=2.0.0
h3==4.0.0b2
gcsfs>=2023.4.0
pydeck>=0.8.0
pydeck-carto>=0.1.0
carto-auth>=0.1.0
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.9
pysal>=23.1
rasterio>=1.3.6
h3ronpy>=0.16.1
geopy>=2.3.0
autopep8>=2.0.2
jupyter-contrib-nbextensions>=0.7.0
notebook>=6.5.4
keplergl>=0.3.2
Jinja2>=3.1.2
geoalchemy2>=0.13.1
